Updates based on feedback
