+++ libs/components/core/src/lib/modules/overlay/overlay-adapter.service.ts
@@ -1,4 +1,13 @@
-import { Injectable, Renderer2, RendererFactory2 } from '@angular/core';
+import { DOCUMENT } from '@angular/common';
+import {
+  Injectable,
+  OnDestroy,
+  Renderer2,
+  RendererFactory2,
+  inject,
+} from '@angular/core';
+
+import { Subscription, fromEvent } from 'rxjs';
 
 /**
  * @internal
@@ -6,13 +15,23 @@ import { Injectable, Renderer2, RendererFactory2 } from '@angular/core';
 @Injectable({
   providedIn: 'root',
 })
-export class SkyOverlayAdapterService {
-  #renderer: Renderer2;
+export class SkyOverlayAdapterService implements OnDestroy {
+  readonly #doc = inject(DOCUMENT);
+
+  readonly #renderer: Renderer2;
+
+  readonly #subscription = new Subscription();
 
   #styleElement: HTMLStyleElement | undefined;
 
   constructor(rendererFactory: RendererFactory2) {
     this.#renderer = rendererFactory.createRenderer(undefined, null);
+    this.#maintainVhProperty();
+  }
+
+  public ngOnDestroy(): void {
+    this.#subscription.unsubscribe();
+    this.#doc.documentElement.style.removeProperty('--skyux-vh');
   }
 
   public restrictBodyScroll(): void {
@@ -39,6 +58,27 @@ export class SkyOverlayAdapterService {
 
   public releaseBodyScroll(): void {
     this.#destroyStyleElement();
+  }
+
+  #maintainVhProperty(): void {
+    // Maintain a --skyux-vh CSS property that is equal to 1% of the viewport height.
+    // Safari on iOS does not maintain accurate vh units when e.g. the keyboard is shown.
+    if (this.#doc.defaultView) {
+      const setVhProperty = (): void => {
+        this.#doc.documentElement.style.setProperty(
+          '--skyux-vh',
+          // Fallback value b/c it can technically be undefined.
+          /* istanbul ignore next */
+          this.#doc.defaultView?.innerHeight
+            ? `${this.#doc.defaultView.innerHeight * 0.01}px`
+            : '1vh'
+        );
+      };
+      setVhProperty();
+      this.#subscription.add(
+        fromEvent(this.#doc.defaultView, 'resize').subscribe(setVhProperty)
+      );
+    }
   }
 
   #destroyStyleElement(): void {
