+++ libs/components/core/src/lib/modules/overlay/overlay.service.ts
@@ -1,4 +1,3 @@
-import { DOCUMENT } from '@angular/common';
 import {
   ApplicationRef,
   ComponentFactoryResolver,
@@ -6,11 +5,7 @@ import {
   EmbeddedViewRef,
   Injectable,
   Injector,
-  OnDestroy,
-  inject,
 } from '@angular/core';
-
-import { Subscription, fromEvent } from 'rxjs';
 
 import { SkyOverlayAdapterService } from './overlay-adapter.service';
 import { SkyOverlayConfig } from './overlay-config';
@@ -25,7 +20,7 @@ import { SkyOverlayComponent } from './overlay.component';
 @Injectable({
   providedIn: 'root',
 })
-export class SkyOverlayService implements OnDestroy {
+export class SkyOverlayService {
   private static overlays: SkyOverlayInstance[] = [];
 
   #adapter: SkyOverlayAdapterService;
@@ -34,11 +29,7 @@ export class SkyOverlayService implements OnDestroy {
 
   #componentFactoryResolver: ComponentFactoryResolver;
 
-  #doc = inject(DOCUMENT);
-
   #injector: Injector;
-
-  #subscription = new Subscription();
 
   // TODO: Replace deprecated `ComponentFactoryResolver`.
   constructor(
@@ -51,12 +42,6 @@ export class SkyOverlayService implements OnDestroy {
     this.#componentFactoryResolver = componentFactoryResolver;
     this.#injector = injector;
     this.#adapter = adapter;
-    this.#maintainVhProperty();
-  }
-
-  public ngOnDestroy(): void {
-    this.#subscription.unsubscribe();
-    this.#doc.documentElement.style.removeProperty('--skyux-vh');
   }
 
   /**
@@ -139,7 +124,7 @@ export class SkyOverlayService implements OnDestroy {
     const domElem = (componentRef.hostView as EmbeddedViewRef<any>)
       .rootNodes[0] as HTMLElement;
 
-    this.#doc.body.appendChild(domElem);
+    document.body.appendChild(domElem);
 
     return componentRef;
   }
@@ -171,26 +156,6 @@ export class SkyOverlayService implements OnDestroy {
       if (!anotherOverlayDisablesScroll) {
         this.#adapter.releaseBodyScroll();
       }
-    }
-  }
-
-  #maintainVhProperty(): void {
-    // Maintain a --skyux-vh CSS property that is equal to 1% of the viewport height.
-    if (this.#doc.defaultView) {
-      const setVhProperty = (): void => {
-        this.#doc.documentElement.style.setProperty(
-          '--skyux-vh',
-          // Fallback value b/c it can technically be undefined.
-          /* istanbul ignore next */
-          this.#doc.defaultView?.innerHeight
-            ? `${this.#doc.defaultView.innerHeight * 0.01}px`
-            : '1vh'
-        );
-      };
-      setVhProperty();
-      this.#subscription.add(
-        fromEvent(this.#doc.defaultView, 'resize').subscribe(setVhProperty)
-      );
     }
   }
 }
