+++ libs/components/theme/src/lib/styles/_modals.scss
@@ -1,5 +1,4 @@
 .sky-modal-body-open {
-  height: 100%;
   overflow: hidden;
   position: relative;
 }
